Refactor WSDL and XSD
